+++ okitweb/static/okit/js/okit_console.js
@@ -5,8 +5,8 @@
 console.debug('Loaded Console Javascript');
 
 
-const okitVersion = '0.44.0';
-const okitReleaseDate = '30th November 2022';
+const okitVersion = '0.45.0';
+const okitReleaseDate = '25th January 2023';
 
 // Validation
 const validate_error_colour = "#ff4d4d";
@@ -120,8 +120,8 @@ function handleConfigChanged(event) {
     console_header_region_select.append($('<option>').attr('value', '').text('Retrieving....'))
     okitRegions.clearLocalStorage()
     okitRegions.load(okitSettings.profile)
-    okitOciData.load(okitSettings.profile, okitSettings.region)
-    setOCILink()
+    // okitOciData.load(okitSettings.profile, okitSettings.region)
+    // setOCILink()
 }
 
 function loadHeaderRegionsDropDown() {
@@ -136,13 +136,14 @@ function loadHeaderRegionsDropDown() {
         okitSettings.save()
     }
     console.info('Region:', okitSettings.region)
-    console_header_region_select.val(okitSettings.region)
-    setOCILink()
+    console_header_region_select.val(okitSettings.region).change()
+    // okitOciData.load(okitSettings.profile, okitSettings.region)
+    // setOCILink()
 }
 
 function handleRegionChanged(event) {
     event = event || window.event;
-    event.stopPropagation()
+    if (event) event.stopPropagation()
     okitSettings.region = $('#console_header_region_select').val()
     okitSettings.save()
     okitOciData.load(okitSettings.profile, okitSettings.region)