+++ visualiser/generators/okitGenerator.py
@@ -27,7 +27,7 @@ from model.okitValidation import OCIJsonValidator
 logger = getLogger()
 
 class OCIGenerator(object):
-    OKIT_VERSION = "0.44.0"
+    OKIT_VERSION = "0.45.0"
     def __init__(self, template_dir, output_dir, visualiser_json, use_vars=False, add_provider=True):
         # Initialise generator output data variables
         self.rendered_resources = {}
@@ -178,6 +178,7 @@ class OCIGenerator(object):
         "keys": "identity",
         "users": "identity",
         "vaults": "identity",
+        "vault_secrets": "identity",
 
         "dhcp_options": "networking",
         "drg_attachments": "networking",